Now Systems class will notify pools when loop ends
Pool internal entity cache now works storing destroyed entities in a new
temporally stack for the next frame so the other systems can check the
destroyed entity (otherwise maybe the entity was reused to instantiate
another !). At the end of the frame (or loop), "Systems" class call the
new 'EndLoop' function in each pool registered and autodestroys all the
entities that were flagged with the new entity function "Destroy()".